requirements.txt
numpy~=1.24.2
Pillow~=9.5.0
tqdm~=4.65.0
torch~=2.0.0
torchvision~=0.15.1
matplotlib~=3.7.1
scikit-learn~=1.2.2
opencv-python~=4.7.0.72